requirements.txt
streamlit
google-generativeai
python-dotenv
PyMuPDF
chromadb
fpdf
pandas
gTTS
streamlit-mic-recorder